Move validation functions to the validation service
