+++ tests/unit_tests/routers/ingest/test_bulk_ingest.py
@@ -14,7 +14,7 @@ from fastapi import status
 from fastapi.testclient import TestClient
 
 from app.errors import ValidationError
-from app.service.ingest import validate_entity_relationships
+from app.service.validation import validate_entity_relationships
 
 
 def test_ingest_when_not_authenticated(client: TestClient):
